restructure notes page
