+++ components/FavoriteButton/index.js
@@ -1,7 +1,7 @@
 import Icon from "@mdi/react";
 import { mdiHeartOutline } from "@mdi/js";
 import { mdiCardsHeart } from "@mdi/js";
-import { styled } from "styled-components";
+import { StyledButton } from "./styles";
 
 export default function FavoriteButton({ isFavored, onClick }) {
   return (
@@ -16,8 +16,3 @@ export default function FavoriteButton({ isFavored, onClick }) {
     </>
   );
 }
-
-const StyledButton = styled.button`
-  border: none;
-  background-color: transparent;
-`;
