+++ components/FavoriteButton/styles.js
@@ -1,0 +1,6 @@
+import { styled } from "styled-components";
+
+export const StyledButton = styled.button`
+  border: none;
+  background-color: transparent;
+`;
