+++ components/Navigation/index.js
@@ -1,18 +1,16 @@
-import Link from "next/link";
-import styled from "styled-components";
-import Icon from "@mdi/react";
 import { mdiMagnify } from "@mdi/js";
 import { mdiNotebookHeartOutline } from "@mdi/js";
 import { mdiCardsHeart } from "@mdi/js";
 import { mdiShimmer } from "@mdi/js";
 import { useRouter } from "next/router";
+import { StyledNavBar, StyledNavBarLink, StyledNavBarIcon } from "./styles";
 
 export default function Navigation() {
   const router = useRouter();
   return (
     <>
-      <NavBar>
-        <StyledNavBarLink href={`/`} active={router.pathname === "/"}>
+      <StyledNavBar>
+        <StyledNavBarLink href={`/`}>
           <StyledNavBarIcon
             path={mdiShimmer}
             size={2}
@@ -20,10 +18,7 @@ export default function Navigation() {
           />
         </StyledNavBarLink>
 
-        <StyledNavBarLink
-          href={`/artworks-collection`}
-          active={router.pathname === "/artworks-collection"}
-        >
+        <StyledNavBarLink href={`/artworks-collection`}>
           <StyledNavBarIcon
             path={mdiMagnify}
             size={2}
@@ -31,50 +26,21 @@ export default function Navigation() {
           />
         </StyledNavBarLink>
 
-        <StyledNavBarLink
-          href={`/my-notes`}
-          active={router.pathname === "/my-notes"}
-        >
+        <StyledNavBarLink href={`/my-notes`}>
           <StyledNavBarIcon
             path={mdiNotebookHeartOutline}
             size={2}
             active={router.pathname === "/my-notes"}
           />
         </StyledNavBarLink>
-        <StyledNavBarLink
-          href={`/favorites`}
-          active={router.pathname === "/favorites"}
-        >
+        <StyledNavBarLink href={`/favorites`}>
           <StyledNavBarIcon
             path={mdiCardsHeart}
             size={2}
             active={router.pathname === "/favorites"}
           />
         </StyledNavBarLink>
-      </NavBar>
+      </StyledNavBar>
     </>
   );
 }
-
-const NavBar = styled.div`
-  position: fixed;
-  bottom: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  height: 60px;
-  background-color: black;
-`;
-
-const StyledNavBarLink = styled(Link)`
-  width: 25%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const StyledNavBarIcon = styled(Icon)`
-  color: ${({ active }) => (active ? "hotpink" : "whitesmoke")};
-`;
